Move Pharos::Kube::Client to separate file
